feat(accepted): Convert complex celebration page to SSG ✅
PHASE 1 - CONVERSION 3/4 COMPLETE - THE COMPLEX ONE!

This was the most challenging conversion:
- Extract massive interactive logic to AcceptedPageClient.tsx (496→415 lines)
- Complex features: confetti animations, window resize handling, search params
- Dynamic share links, clipboard API, audio controls, Suspense boundaries
- All preserved: URL params (child, program, location), personalized welcome

Results:
✅ Build successful - no RSC/hook/Suspense errors
✅ Static HTML generated (.next/server/app/accepted.html - 17.0kB)
✅ SEO meta: 'Welcome to Teddy Kids! Your child has been accepted...'
✅ Bundle performance stable

🎊 Complex interactive celebration page now statically rendered!
Next: /about/policy (final conversion) → PHASE 1 COMPLETE!
